Replace failure with thiserror Refactor main module into submodules
